Move LeafCohortParameters class from Leaf.cs into its own file
